+++ src/core/converter/ImportFileAeroflyCustomMissionsJsonConverter.test.ts
@@ -1,12 +1,16 @@
 import { describe, it } from "node:test";
 import assert from "node:assert";
 import fs from "node:fs";
-import { AeroflyFlightFallback } from "../util/AeroflyFlightFallback.js";
+import path from "node:path";
+import { AeroflyFlightFallback } from "../data/AeroflyFlightFallback.js";
 import { ImportFileAeroflyCustomMissionsJsonConverter } from "./ImportFileAeroflyCustomMissionsJsonConverter.js";
 
 describe("ImportFileAeroflyCustomMissionsJsonConverter", () => {
     it("should find multiple flight plans in a .aerofly.json file", () => {
-        const content = fs.readFileSync("./src/test/fixtures/matthias-rust.aerofly.json", "utf-8");
+        const content = fs.readFileSync(
+            path.join(import.meta.dirname, "../../..", "src/test/fixtures", "matthias-rust.aerofly.json"),
+            "utf-8",
+        );
 
         const converter = new ImportFileAeroflyCustomMissionsJsonConverter();
         const indices = converter.getIndices(content);
@@ -18,7 +22,10 @@ describe("ImportFileAeroflyCustomMissionsJsonConverter", () => {
     it("should convert .aerofly.json files into AeroflyFlight classes", () => {
         const flight = new AeroflyFlightFallback();
 
-        const content = fs.readFileSync("./src/test/fixtures/matthias-rust.aerofly.json", "utf-8");
+        const content = fs.readFileSync(
+            path.join(import.meta.dirname, "../../..", "src/test/fixtures", "matthias-rust.aerofly.json"),
+            "utf-8",
+        );
 
         const converter = new ImportFileAeroflyCustomMissionsJsonConverter();
         converter.convert(content, flight);